0.1 components operational
