+++ contract-env/cu/src/routes/base.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 router.get('', async (req, res) => {
-    res.send(`ao relay node`);
+    res.send(`ao compute unit`);
 });
 
 module.exports = router;
