+++ contract-env/mu/.gitignore
@@ -1,0 +1,10 @@
+**/node_modules
+**/bin
+**/build
+**/wallets
+**/cache
+**/sindex
+**/.DS_Store
+**/data
+*yalc*
+ao-cache*
