+++ contract-env/mu/src/routes/base.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 router.get('', async (req, res) => {
-    res.send(`ao dre node`);
+    res.send(`ao messenger unit`);
 });
 
 module.exports = router;
