--- contract-env/relay/.dockerignore
@@ -1,2 +1,0 @@
-node_modules/
-npm-debug.log
