--- contract-env/relay/.gitignore
@@ -1,10 +1,0 @@
-**/node_modules
-**/bin
-**/build
-**/wallets
-**/cache
-**/sindex
-**/.DS_Store
-**/data
-*yalc*
-ao-cache*
