--- contract-env/relay/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "relay",
-    "version": "1.0.0",
-    "description": "A basic Express.js application",
-    "main": "src/app.js",
-    "scripts": {
-        "start": "node src/app.js",
-        "start:local": "nodemon src/app.js",
-        "docker-build": "docker build -t relay-app .",
-        "docker-run": "npm run docker-build && docker run -p 3004:3004 relay-app"
-    },
-    "dependencies": {
-        "@permaweb/ao-sdk": "file:../../sdk",
-        "body-parser": "^1.20.2",
-        "cors": "^2.8.5",
-        "express": "^4.17.1"
-    },
-    "devDependencies": {
-        "nodemon": "^3.0.1"
-    },
-    "keywords": [
-        "express",
-        "web",
-        "server"
-    ],
-    "author": "VinceJuliano",
-    "license": "MIT"
-}
