--- contract-env/relay/src/config.js
@@ -1,11 +1,0 @@
-const path = require('path');
-const fs = require('fs');
-
-const walletPath = process.env.PATH_TO_WALLET;
-let walletKey = JSON.parse(fs.readFileSync(path.resolve(walletPath), 'utf8'));
-
-let config = {
-    relayWallet: walletKey
-};
-
-module.exports = config;
